+++ .gitignore
@@ -7,17 +7,87 @@
 .vscode/solution-explorer/interface.ts-template
 .vscode/solution-explorer/template-list.json
 .vscode/solution-explorer/template-parameters.js
-ApiMentoria/bin/**
-ApiMentoria/obj/**
-Service/bin/**
-Service/obj/**
-Core/obj/**
-Repository/obj/**
-Tests/bin/**
-Tests/obj/**
+ApiMentoria/bin/*
+ApiMentoria/obj/*
+Service/bin/*
+Service/obj/*
+Core/obj/*
+Repository/obj/*
+Tests/bin/*
+Tests/obj/*
 ApiMentoria/obj/Debug/netcoreapp3.1/ApiMentoria.assets.cache
 Repository/bin/Debug/netcoreapp3.1/Core.dll
 Repository/bin/Debug/netcoreapp3.1/Core.pdb
+Repository/bin/Debug/netcoreapp3.1/Repository.dll
+Repository/bin/Debug/netcoreapp3.1/Repository.pdb
+Repository/bin/Debug/netcoreapp3.1/Repository.runtimeconfig.dev.json
+Repository/obj/project.assets.json
+Repository/obj/Repository.csproj.nuget.dgspec.json
+Repository/obj/Repository.csproj.nuget.g.props
+Repository/obj/Debug/netcoreapp3.1/Repository.assets.cache
+Repository/obj/Debug/netcoreapp3.1/Repository.csproj.FileListAbsolute.txt
+Repository/obj/Debug/netcoreapp3.1/Repository.csprojAssemblyReference.cache
+Repository/obj/Debug/netcoreapp3.1/Repository.dll
+Repository/obj/Debug/netcoreapp3.1/Repository.pdb
+Repository/obj/Debug/netcoreapp3.1/Repository.assets.cache
+ApiMentoria/bin/Debug/netcoreapp3.1/ApiMentoria.deps.json
+ApiMentoria/bin/Debug/netcoreapp3.1/ApiMentoria.dll
+ApiMentoria/bin/Debug/netcoreapp3.1/ApiMentoria.pdb
+ApiMentoria/bin/Debug/netcoreapp3.1/ApiMentoria.runtimeconfig.dev.json
+ApiMentoria/obj/ApiMentoria.csproj.nuget.cache
+ApiMentoria/obj/ApiMentoria.csproj.nuget.dgspec.json
+ApiMentoria/obj/ApiMentoria.csproj.nuget.g.props
+ApiMentoria/obj/project.assets.json
+ApiMentoria/obj/Debug/netcoreapp3.1/ApiMentoria.assets.cache
+ApiMentoria/obj/Debug/netcoreapp3.1/ApiMentoria.csproj.FileListAbsolute.txt
+ApiMentoria/obj/Debug/netcoreapp3.1/ApiMentoria.csprojAssemblyReference.cache
+ApiMentoria/obj/Debug/netcoreapp3.1/ApiMentoria.dll
+ApiMentoria/obj/Debug/netcoreapp3.1/ApiMentoria.pdb
+Core/bin/Debug/netcoreapp3.1/Core.dll
+Core/bin/Debug/netcoreapp3.1/Core.pdb
+Core/obj/Core.csproj.nuget.dgspec.json
+Core/obj/Core.csproj.nuget.g.props
+Core/obj/project.assets.json
+Core/obj/Debug/netcoreapp3.1/Core.assets.cache
+Core/obj/Debug/netcoreapp3.1/Core.csproj.FileListAbsolute.txt
+Core/obj/Debug/netcoreapp3.1/Core.csprojAssemblyReference.cache
+Core/obj/Debug/netcoreapp3.1/Core.dll
+Core/obj/Debug/netcoreapp3.1/Core.pdb
+ApiMentoria/obj/Debug/netcoreapp3.1/ApiMentoria.assets.cache
+Core/obj/Debug/netcoreapp3.1/Core.assets.cache
+Infrastructure/bin/Debug/netcoreapp3.1/Core.dll
+Infrastructure/bin/Debug/netcoreapp3.1/Core.pdb
+Infrastructure/bin/Debug/netcoreapp3.1/Infrastructure.deps.json
+Infrastructure/bin/Debug/netcoreapp3.1/Infrastructure.dll
+Infrastructure/bin/Debug/netcoreapp3.1/Infrastructure.pdb
+Infrastructure/bin/Debug/netstandard2.0/Infrastructure.deps.json
+Infrastructure/bin/Debug/netstandard2.0/Infrastructure.dll
+Infrastructure/bin/Debug/netstandard2.0/Infrastructure.pdb
+Infrastructure/obj/Infrastructure.csproj.nuget.cache
+Infrastructure/obj/Infrastructure.csproj.nuget.dgspec.json
+Infrastructure/obj/Infrastructure.csproj.nuget.g.props
+Infrastructure/obj/Infrastructure.csproj.nuget.g.targets
+Infrastructure/obj/project.assets.json
+Infrastructure/obj/Debug/netcoreapp3.1/.NETCoreApp,Version=v3.1.AssemblyAttributes.cs
+Infrastructure/obj/Debug/netcoreapp3.1/Infrastructure.AssemblyInfo.cs
+Infrastructure/obj/Debug/netcoreapp3.1/Infrastructure.AssemblyInfoInputs.cache
+Infrastructure/obj/Debug/netcoreapp3.1/Infrastructure.assets.cache
+Infrastructure/obj/Debug/netcoreapp3.1/Infrastructure.csproj.CopyComplete
+Infrastructure/obj/Debug/netcoreapp3.1/Infrastructure.csproj.FileListAbsolute.txt
+Infrastructure/obj/Debug/netcoreapp3.1/Infrastructure.csprojAssemblyReference.cache
+Infrastructure/obj/Debug/netcoreapp3.1/Infrastructure.dll
+Infrastructure/obj/Debug/netcoreapp3.1/Infrastructure.pdb
+Infrastructure/obj/Debug/netstandard2.0/.NETStandard,Version=v2.0.AssemblyAttributes.cs
+Infrastructure/obj/Debug/netstandard2.0/Infrastructure.AssemblyInfo.cs
+Infrastructure/obj/Debug/netstandard2.0/Infrastructure.AssemblyInfoInputs.cache
+Infrastructure/obj/Debug/netstandard2.0/Infrastructure.assets.cache
+Infrastructure/obj/Debug/netstandard2.0/Infrastructure.csproj.FileListAbsolute.txt
+Infrastructure/obj/Debug/netstandard2.0/Infrastructure.csprojAssemblyReference.cache
+Infrastructure/obj/Debug/netstandard2.0/Infrastructure.dll
+Infrastructure/obj/Debug/netstandard2.0/Infrastructure.pdb
+Repository/bin/Debug/netcoreapp3.1/Core.dll
+Repository/bin/Debug/netcoreapp3.1/Core.pdb
+Repository/bin/Debug/netcoreapp3.1/Repository.deps.json
 Repository/bin/Debug/netcoreapp3.1/Repository.dll
 Repository/bin/Debug/netcoreapp3.1/Repository.pdb
 Repository/bin/Debug/netcoreapp3.1/Repository.runtimeconfig.dev.json
