+++ Core/Entities/BaseEntity.cs
@@ -5,7 +5,7 @@ namespace Core.Entities
     public abstract class BaseEntity
     {
         [Key]
-        public int Id { get; set; }
+        public virtual int Id { get; set; }
 
         public BaseEntity() { }
     }
